+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Add unbalanced parentheses/EOF test
+
 # 0.12.1.0
 
 - GHC 8.4.1 support